requirements.txt
streamlit==1.17.0
tensorflow==2.13.0
pandas==1.5.3
scikit-learn==1.2.2
Pillow==9.5.0
numpy==1.23.5